--- src/main/java/com/example/jobgsm/domain/email/domain/exception/MisMatchAuthCodeException.java
@@ -1,15 +1,0 @@
-package com.example.jobgsm.domain.email.domain.exception;
-
-import com.example.jobgsm.domain.email.global.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class MisMatchAuthCodeException extends RuntimeException{
-
-    private final ErrorCode errorCode;
-
-    public MisMatchAuthCodeException(String message) {
-        super(message);
-        this.errorCode = ErrorCode.MISMATCH_AUTH_CODE;
-    }
-}
