+++ src/main/resources/application.properties
@@ -5,13 +5,9 @@ spring.redis.port=6379
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=jobGSM2022@gmail.com
-spring.mail.password=yotqfsozeycdhxqb
-spring.mail.default-encoding=utf-8
+spring.mail.username=jobGSM2023@gmail.com
+spring.mail.password=yvxcaenfpjfdffkw
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
-spring.mail.properties.mail.smtp.starttls.required=true
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
-\
-
